Address packing for Shelley
