+++ crypto-primitives.js
@@ -1,0 +1,182 @@
+const Module = require('./lib.js')
+
+const {validateBuffer, validateArray} = require("./utils/validation")
+
+function blake2b(input, outputLen) {
+  validateBuffer(input)
+
+  const inputLen = input.length
+  const inputArrPtr = Module._malloc(inputLen)
+  const inputArr = new Uint8Array(Module.HEAPU8.buffer, inputArrPtr, inputLen)
+  const outputArrPtr = Module._malloc(outputLen)
+  const outputArr = new Uint8Array(Module.HEAPU8.buffer, outputArrPtr, outputLen)
+
+  inputArr.set(input)
+
+  Module._emscripten_blake2b(inputArrPtr, inputLen, outputArrPtr, outputLen)
+
+  Module._free(inputArrPtr)
+  Module._free(outputArrPtr)
+
+  return Buffer.from(outputArr)
+}
+
+function hmac_sha512(initKey, inputs) {
+  validateBuffer(initKey)
+  validateArray(inputs)
+  inputs.map(validateBuffer)
+
+  const ctxLen = Module._emscripten_size_of_hmac_sha512_ctx()
+  const ctxArrPtr = Module._malloc(ctxLen)
+  const ctxArr = new Uint8Array(Module.HEAPU8.buffer, ctxArrPtr, ctxLen)
+
+  const initKeyLen = initKey.length
+  const initKeyArrPtr = Module._malloc(initKeyLen)
+  const initKeyArr = new Uint8Array(Module.HEAPU8.buffer, initKeyArrPtr, initKeyLen)
+
+  const outputLen = 64
+  const outputArrPtr = Module._malloc(outputLen)
+  const outputArr = new Uint8Array(Module.HEAPU8.buffer, outputArrPtr, outputLen)
+
+  initKeyArr.set(initKey)
+
+  Module._emscripten_hmac_sha512_init(ctxArrPtr, initKeyArrPtr, initKeyLen)
+
+  for (let i = 0; i < inputs.length; i++) {
+    const inputLen = inputs[i].length
+    const inputArrPtr = Module._malloc(inputLen)
+    const inputArr = new Uint8Array(Module.HEAPU8.buffer, inputArrPtr, inputLen)
+
+    inputArr.set(inputs[i])
+
+    Module._emscripten_hmac_sha512_update(ctxArrPtr, inputArrPtr, inputLen)
+
+    Module._free(inputArrPtr)
+  }
+
+  Module._emscripten_hmac_sha512_final(ctxArrPtr, outputArrPtr)
+
+  Module._free(initKeyArrPtr)
+  Module._free(ctxArrPtr)
+  Module._free(outputArrPtr)
+
+  return Buffer.from(outputArr)
+}
+
+
+function chacha20poly1305Encrypt(input, key, nonce) {
+  validateBuffer(input)
+  validateBuffer(key, 32)
+  validateBuffer(nonce, 12)
+
+  const inputLen = input.length
+  const inputArrPtr = Module._malloc(inputLen)
+  const inputArr = new Uint8Array(Module.HEAPU8.buffer, inputArrPtr, inputLen)
+
+  const keyLen = key.length
+  const keyArrPtr = Module._malloc(keyLen)
+  const keyArr = new Uint8Array(Module.HEAPU8.buffer, keyArrPtr, keyLen)
+
+  const nonceLen = nonce.length
+  const nonceArrPtr = Module._malloc(nonceLen)
+  const nonceArr = new Uint8Array(Module.HEAPU8.buffer, nonceArrPtr, nonceLen)
+
+  const tagLen = 16
+  const outputLen = inputLen + tagLen
+  const outputArrPtr = Module._malloc(outputLen)
+  const outputArr = new Uint8Array(Module.HEAPU8.buffer, outputArrPtr, outputLen)
+
+  inputArr.set(input)
+  keyArr.set(key)
+  nonceArr.set(nonce)
+
+  const resultCode = Module._emscripten_chacha20poly1305_enc(keyArrPtr, nonceArrPtr, inputArrPtr, inputLen, outputArrPtr, outputArrPtr + inputLen, tagLen, 1)
+
+  Module._free(inputArrPtr)
+  Module._free(keyArrPtr)
+  Module._free(nonceArrPtr)
+  Module._free(outputArrPtr)
+
+  if (resultCode !== 0) {
+    throw Error('chacha20poly1305 encryption has failed!')
+  }
+
+  return Buffer.from(outputArr)
+}
+
+function chacha20poly1305Decrypt(input, key, nonce) {
+  validateBuffer(input)
+  validateBuffer(key, 32)
+  validateBuffer(nonce, 12)
+
+  // extract tag from input
+  const tagLen = 16
+  const tag = input.slice(input.length - tagLen, input.length)
+  input = input.slice(0, input.length - tagLen)
+
+  const inputLen = input.length
+  const inputArrPtr = Module._malloc(inputLen)
+  const inputArr = new Uint8Array(Module.HEAPU8.buffer, inputArrPtr, inputLen)
+
+  const tagArrPtr = Module._malloc(tagLen)
+  const tagArr = new Uint8Array(Module.HEAPU8.buffer, tagArrPtr, tagLen)
+
+  const keyLen = key.length
+  const keyArrPtr = Module._malloc(keyLen)
+  const keyArr = new Uint8Array(Module.HEAPU8.buffer, keyArrPtr, keyLen)
+
+  const nonceLen = nonce.length
+  const nonceArrPtr = Module._malloc(nonceLen)
+  const nonceArr = new Uint8Array(Module.HEAPU8.buffer, nonceArrPtr, nonceLen)
+
+  const outputLen = inputLen
+  const outputArrPtr = Module._malloc(outputLen)
+  const outputArr = new Uint8Array(Module.HEAPU8.buffer, outputArrPtr, outputLen)
+
+  inputArr.set(input)
+  tagArr.set(tag)
+  keyArr.set(key)
+  nonceArr.set(nonce)
+
+  const resultCode = Module._emscripten_chacha20poly1305_enc(keyArrPtr, nonceArrPtr, inputArrPtr, inputLen, outputArrPtr, tagArrPtr, tagLen, 0)
+
+  Module._free(inputArrPtr)
+  Module._free(keyArrPtr)
+  Module._free(nonceArrPtr)
+  Module._free(outputArrPtr)
+  Module._free(tagArrPtr)
+
+  if (resultCode !== 0) {
+    throw Error('chacha20poly1305 decryption has failed!')
+  }
+
+  return Buffer.from(outputArr)
+}
+
+function sha3_256(input) {
+  validateBuffer(input)
+  const inputLen = input.length
+  const inputArrPtr = Module._malloc(inputLen)
+  const inputArr = new Uint8Array(Module.HEAPU8.buffer, inputArrPtr, inputLen)
+
+  const outputLen = 32
+  const outputArrPtr = Module._malloc(outputLen)
+  const outputArr = new Uint8Array(Module.HEAPU8.buffer, outputArrPtr, outputLen)
+
+  inputArr.set(input)
+
+  Module._emscripten_sha3_256(inputArrPtr, inputLen, outputArrPtr)
+
+  Module._free(inputArrPtr)
+  Module._free(outputArrPtr)
+
+  return Buffer.from(outputArr)
+}
+
+module.exports = {
+  blake2b,
+  chacha20poly1305Decrypt,
+  chacha20poly1305Encrypt,
+  hmac_sha512,
+  sha3_256
+}
