+++ index.js
@@ -1,81 +1,16 @@
+const {validateBuffer, validateDerivationIndex, validateDerivationScheme, validateMnemonic, validateArray, validateString, validatePaperWalletMnemonic} = require("./utils/validation")
+const address = require("./address");
+const crypto = require("./crypto-primitives")
+
+const bech32 = require("bech32")
 const bip39 = require('bip39')
-const cbor = require('borc')
 
 const Module = require('./lib.js')
-const crc32 = require('./utils/crc32')
 const base58 = require('./utils/base58')
 const scrypt = require('./utils/scrypt-async')
 const pbkdf2 = require('./utils/pbkdf2')
-const CborIndefiniteLengthArray = require('./utils/CborIndefiniteLengthArray')
 
 const HARDENED_THRESHOLD = 0x80000000
-
-function validateDerivationScheme(input) {
-  if (input !== 1 && input !== 2) {
-    throw new Error('invalid derivation scheme!')
-  }
-}
-
-function validateBuffer(input, expectedLength) {
-  if (!Buffer.isBuffer(input)) {
-    throw new Error('not buffer!')
-  }
-
-  if (expectedLength && input.length !== expectedLength) {
-    throw new Error('Invalid buffer length')
-  }
-}
-
-function validateArray(input) {
-  if (typeof input !== typeof []) {
-    throw new Error('not an array!')
-  }
-}
-
-function validateDerivationIndex(input) {
-  if (!Number.isInteger(input)) {
-    throw new Error('invalid derivation index!')
-  }
-}
-
-function validateString(input) {
-  if (typeof input !== typeof 'aa') {
-    throw new Error('not a string!')
-  }
-}
-
-function validateMnemonic(input) {
-  if (!bip39.validateMnemonic(input)) {
-    const e = new Error('Invalid or unsupported mnemonic format:')
-    e.name = 'InvalidArgumentException'
-    throw e
-  }
-}
-
-function validateMnemonicWords(input) {
-  const wordlist = bip39.wordlists.EN
-  const words = input.split(' ')
-
-  const valid = words.reduce((result, word) => {
-    return result && wordlist.indexOf(word) !== -1
-  }, true)
-
-  if (!valid) {
-    throw new Error('Invalid mnemonic words')
-  }
-}
-
-function validatePaperWalletMnemonic(input) {
-  validateMnemonicWords(input)
-
-  const mnemonicLength = input.split(' ').length
-
-  if (mnemonicLength !== 27) {
-    throw Error(
-      `Paper Wallet Mnemonic must be 27 words, got ${mnemonicLength} instead`
-    )
-  }
-}
 
 function cborEncodeBuffer(input) {
   validateBuffer(input)
@@ -163,15 +98,15 @@ function mnemonicToRootKeypairV1(mnemonic) {
 function mnemonicToSeedV1(mnemonic) {
   validateMnemonic(mnemonic)
   const entropy = Buffer.from(bip39.mnemonicToEntropy(mnemonic), 'hex')
-  
-  return cborEncodeBuffer(blake2b(cborEncodeBuffer(entropy), 32))
+
+  return cborEncodeBuffer(crypto.blake2b(cborEncodeBuffer(entropy), 32))
 }
 
 function seedToKeypairV1(seed) {
   let result
   for (let i = 1; result === undefined && i <= 1000; i++) {
     try {
-      const digest = hmac_sha512(seed, [Buffer.from(`Root Seed Chain ${i}`, 'ascii')])
+      const digest = crypto.hmac_sha512(seed, [Buffer.from(`Root Seed Chain ${i}`, 'ascii')])
       const tempSeed = digest.slice(0, 32)
       const chainCode = digest.slice(32, 64)
 
@@ -321,87 +256,6 @@ function derivePublic(parentExtPubKey, index, derivationScheme) {
   return Buffer.concat([Buffer.from(childPubKeyArr), Buffer.from(childChainCodeArr)])
 }
 
-function blake2b(input, outputLen) {
-  validateBuffer(input)
-
-  const inputLen = input.length
-  const inputArrPtr = Module._malloc(inputLen)
-  const inputArr = new Uint8Array(Module.HEAPU8.buffer, inputArrPtr, inputLen)
-  const outputArrPtr = Module._malloc(outputLen)
-  const outputArr = new Uint8Array(Module.HEAPU8.buffer, outputArrPtr, outputLen)
-
-  inputArr.set(input)
-
-  Module._emscripten_blake2b(inputArrPtr, inputLen, outputArrPtr, outputLen)
-
-  Module._free(inputArrPtr)
-  Module._free(outputArrPtr)
-
-  return Buffer.from(outputArr)
-}
-
-function sha3_256(input) {
-  validateBuffer(input)
-  const inputLen = input.length
-  const inputArrPtr = Module._malloc(inputLen)
-  const inputArr = new Uint8Array(Module.HEAPU8.buffer, inputArrPtr, inputLen)
-
-  const outputLen = 32
-  const outputArrPtr = Module._malloc(outputLen)
-  const outputArr = new Uint8Array(Module.HEAPU8.buffer, outputArrPtr, outputLen)
-
-  inputArr.set(input)
-
-  Module._emscripten_sha3_256(inputArrPtr, inputLen, outputArrPtr)
-
-  Module._free(inputArrPtr)
-  Module._free(outputArrPtr)
-
-  return Buffer.from(outputArr)
-}
-
-function hmac_sha512(initKey, inputs) {
-  validateBuffer(initKey)
-  validateArray(inputs)
-  inputs.map(validateBuffer)
-
-  const ctxLen = Module._emscripten_size_of_hmac_sha512_ctx()
-  const ctxArrPtr = Module._malloc(ctxLen)
-  const ctxArr = new Uint8Array(Module.HEAPU8.buffer, ctxArrPtr, ctxLen)
-
-  const initKeyLen = initKey.length
-  const initKeyArrPtr = Module._malloc(initKeyLen)
-  const initKeyArr = new Uint8Array(Module.HEAPU8.buffer, initKeyArrPtr, initKeyLen)
-
-  const outputLen = 64
-  const outputArrPtr = Module._malloc(outputLen)
-  const outputArr = new Uint8Array(Module.HEAPU8.buffer, outputArrPtr, outputLen)
-
-  initKeyArr.set(initKey)
-
-  Module._emscripten_hmac_sha512_init(ctxArrPtr, initKeyArrPtr, initKeyLen)
-
-  for (let i = 0; i < inputs.length; i++) {
-    const inputLen = inputs[i].length
-    const inputArrPtr = Module._malloc(inputLen)
-    const inputArr = new Uint8Array(Module.HEAPU8.buffer, inputArrPtr, inputLen)
-
-    inputArr.set(inputs[i])
-
-    Module._emscripten_hmac_sha512_update(ctxArrPtr, inputArrPtr, inputLen)
-
-    Module._free(inputArrPtr)
-  }
-
-  Module._emscripten_hmac_sha512_final(ctxArrPtr, outputArrPtr)
-
-  Module._free(initKeyArrPtr)
-  Module._free(ctxArrPtr)
-  Module._free(outputArrPtr)
-
-  return Buffer.from(outputArr)
-}
-
 function cardanoMemoryCombine(input, password) {
   validateString(password)
   validateBuffer(input)
@@ -410,7 +264,7 @@ function cardanoMemoryCombine(input, password) {
     return input
   }
 
-  const transformedPassword = blake2b(Buffer.from(password, 'utf-8'), 32)
+  const transformedPassword = crypto.blake2b(Buffer.from(password, 'utf-8'), 32)
   const transformedPasswordLen = transformedPassword.length
   const transformedPasswordArrPtr = Module._malloc(transformedPasswordLen)
   const transformedPasswordArr = new Uint8Array(Module.HEAPU8.buffer, transformedPasswordArrPtr, transformedPasswordLen)
@@ -430,95 +284,6 @@ function cardanoMemoryCombine(input, password) {
   Module._free(inputArrPtr)
   Module._free(outputArrPtr)
   Module._free(transformedPasswordArrPtr)
-
-  return Buffer.from(outputArr)
-}
-
-function chacha20poly1305Encrypt(input, key, nonce) {
-  validateBuffer(input)
-  validateBuffer(key, 32)
-  validateBuffer(nonce, 12)
-
-  const inputLen = input.length
-  const inputArrPtr = Module._malloc(inputLen)
-  const inputArr = new Uint8Array(Module.HEAPU8.buffer, inputArrPtr, inputLen)
-
-  const keyLen = key.length
-  const keyArrPtr = Module._malloc(keyLen)
-  const keyArr = new Uint8Array(Module.HEAPU8.buffer, keyArrPtr, keyLen)
-
-  const nonceLen = nonce.length
-  const nonceArrPtr = Module._malloc(nonceLen)
-  const nonceArr = new Uint8Array(Module.HEAPU8.buffer, nonceArrPtr, nonceLen)
-
-  const tagLen = 16
-  const outputLen = inputLen + tagLen
-  const outputArrPtr = Module._malloc(outputLen)
-  const outputArr = new Uint8Array(Module.HEAPU8.buffer, outputArrPtr, outputLen)
-
-  inputArr.set(input)
-  keyArr.set(key)
-  nonceArr.set(nonce)
-
-  const resultCode = Module._emscripten_chacha20poly1305_enc(keyArrPtr, nonceArrPtr, inputArrPtr, inputLen, outputArrPtr, outputArrPtr + inputLen, tagLen, 1)
-
-  Module._free(inputArrPtr)
-  Module._free(keyArrPtr)
-  Module._free(nonceArrPtr)
-  Module._free(outputArrPtr)
-
-  if (resultCode !== 0) {
-    throw Error('chacha20poly1305 encryption has failed!')
-  }
-
-  return Buffer.from(outputArr)
-}
-
-function chacha20poly1305Decrypt(input, key, nonce) {
-  validateBuffer(input)
-  validateBuffer(key, 32)
-  validateBuffer(nonce, 12)
-
-  // extract tag from input
-  const tagLen = 16
-  const tag = input.slice(input.length - tagLen, input.length)
-  input = input.slice(0, input.length - tagLen)
-
-  const inputLen = input.length
-  const inputArrPtr = Module._malloc(inputLen)
-  const inputArr = new Uint8Array(Module.HEAPU8.buffer, inputArrPtr, inputLen)
-
-  const tagArrPtr = Module._malloc(tagLen)
-  const tagArr = new Uint8Array(Module.HEAPU8.buffer, tagArrPtr, tagLen)
-
-  const keyLen = key.length
-  const keyArrPtr = Module._malloc(keyLen)
-  const keyArr = new Uint8Array(Module.HEAPU8.buffer, keyArrPtr, keyLen)
-
-  const nonceLen = nonce.length
-  const nonceArrPtr = Module._malloc(nonceLen)
-  const nonceArr = new Uint8Array(Module.HEAPU8.buffer, nonceArrPtr, nonceLen)
-
-  const outputLen = inputLen
-  const outputArrPtr = Module._malloc(outputLen)
-  const outputArr = new Uint8Array(Module.HEAPU8.buffer, outputArrPtr, outputLen)
-
-  inputArr.set(input)
-  tagArr.set(tag)
-  keyArr.set(key)
-  nonceArr.set(nonce)
-
-  const resultCode = Module._emscripten_chacha20poly1305_enc(keyArrPtr, nonceArrPtr, inputArrPtr, inputLen, outputArrPtr, tagArrPtr, tagLen, 0)
-
-  Module._free(inputArrPtr)
-  Module._free(keyArrPtr)
-  Module._free(nonceArrPtr)
-  Module._free(outputArrPtr)
-  Module._free(tagArrPtr)
-
-  if (resultCode !== 0) {
-    throw Error('chacha20poly1305 decryption has failed!')
-  }
 
   return Buffer.from(outputArr)
 }
@@ -567,114 +332,22 @@ async function paperWalletUnscrambleStrings(passphrase, mnemonic) {
 async function xpubToHdPassphrase(xpub) {
   validateBuffer(xpub, 64)
 
-  return pbkdf2(xpub, 'address-hashing', 500, 32, 'sha512')  
+  return pbkdf2(xpub, 'address-hashing', 500, 32, 'sha512')
 }
 
-function packAddress(derivationPath, xpub, hdPassphrase, derivationScheme) {
-  validateBuffer(xpub, 64)
-  validateDerivationScheme(derivationScheme)
-
-  if (derivationScheme === 1) {
-    validateArray(derivationPath)
-    validateBuffer(hdPassphrase, 32)
-  }
-
-  let addressPayload, addressAttributes
-  if (derivationScheme === 1 && derivationPath.length > 0) {
-    addressPayload = encryptDerivationPath(derivationPath, hdPassphrase)
-    addressAttributes = new Map([[1, cbor.encode(addressPayload)]])
-  } else {
-    addressPayload = Buffer.from([])
-    addressAttributes = new Map()
-  }
-
-  const addressRoot = getAddressHash([
-    0,
-    [0, xpub],
-    addressPayload.length > 0 ? new Map([[1, cbor.encode(addressPayload)]]) : new Map(),
-  ])
-  const addressType = 0 // Public key address
-  const addressData = [addressRoot, addressAttributes, addressType]
-  const addressDataEncoded = cbor.encode(addressData)
-
-  return base58.encode(
-    cbor.encode([new cbor.Tagged(24, addressDataEncoded), crc32(addressDataEncoded)])
-  )
+function bech32Encode(prefix, data) {
+  const words = bech32.toWords(data)
+  // we need longer than default length for privkeys and 1000 should suffice
+  return bech32.encode(prefix, words, 1000)
 }
 
-function unpackAddress(address, hdPassphrase) {
-  // we decode the address from the base58 string
-  // and then we strip the 24 CBOR data tags (the "[0].value" part)
-  const addressAsBuffer = cbor.decode(base58.decode(address))[0].value
-  const addressData = cbor.decode(addressAsBuffer)
-  const attributes = addressData[1]
-  const payload = cbor.decode(attributes.get(1))
-  let derivationPath
-
-  try {
-    derivationPath = decryptDerivationPath(payload, hdPassphrase)
-  } catch (e) {
-    throw new Error('Unable to get derivation path from address')
-  }
-
-  if (derivationPath && derivationPath.length > 2) {
-    throw Error('Invalid derivation path length, should be at most 2')
-  }
-
+function bech32Decode(str){
+  const tmp = bech32.decode(str, 1000)
   return {
-    derivationPath,
+    prefix: tmp.prefix,
+    data: Buffer.from(bech32.fromWords(tmp.words)),
   }
 }
-
-function isValidAddress(address) {
-  try {
-    // we decode the address from the base58 string
-    const addressAsArray = cbor.decode(base58.decode(address))
-    // we strip the 24 CBOR data taga by taking the "value" attribute from the "Tagged" object
-    const addressDataEncoded = addressAsArray[0].value
-    const crc32Checksum = addressAsArray[1]
-    
-    if (crc32Checksum !== crc32(addressDataEncoded)) {
-      return false
-    }
-    
-  } catch (e) {
-    return false
-  }
-  return true
-}
-
-
-function getAddressHash(input) {
-  // eslint-disable-next-line camelcase
-  const firstHash = sha3_256(cbor.encode(input))
-  return blake2b(firstHash, 28)
-}
-
-function encryptDerivationPath(derivationPath, hdPassphrase) {
-  const serializedDerivationPath = cbor.encode(new CborIndefiniteLengthArray(derivationPath))
-
-  return chacha20poly1305Encrypt(
-    serializedDerivationPath,
-    hdPassphrase,
-    Buffer.from('serokellfore')
-  )
-}
-
-function decryptDerivationPath(addressPayload, hdPassphrase) {
-  const decipheredDerivationPath = chacha20poly1305Decrypt(
-    addressPayload,
-    hdPassphrase,
-    Buffer.from('serokellfore')
-  )
-
-  try {
-    return cbor.decode(Buffer.from(decipheredDerivationPath))
-  } catch (err) {
-    throw new Error('incorrect address or passphrase')
-  }
-}
-
 
 module.exports = {
   derivePublic,
@@ -684,19 +357,25 @@ module.exports = {
   mnemonicToRootKeypair,
   decodePaperWalletMnemonic,
   xpubToHdPassphrase,
-  packAddress,
-  unpackAddress,
-  isValidAddress,
+  packBootstrapAddress: address.packBootstrapAddress,
+  packBaseAddress: address.packBaseAddress,
+  packPointerAddress: address.packPointerAddress,
+  packEnterpriseAddress: address.packEnterpriseAddress,
+  getAddressInfo: address.getAddressInfo,
+  unpackAddress: address.unpackAddress,
+  isValidAddress: address.isValidAddress,
   cardanoMemoryCombine,
-  blake2b,
+  blake2b: crypto.blake2b,
   base58,
+  bech32_encode: bech32Encode,
+  bech32_decode: bech32Decode,
   scrypt,
   toPublic,
   _mnemonicToSeedV1: mnemonicToSeedV1,
   _seedToKeypairV1: seedToKeypairV1,
   _mnemonicToSeedV2: mnemonicToSeedV2,
   _seedToKeypairV2: seedToKeypairV2,
-  _sha3_256: sha3_256,
-  _chacha20poly1305Decrypt: chacha20poly1305Decrypt,
-  _chacha20poly1305Encrypt: chacha20poly1305Encrypt,
+  _sha3_256: crypto.sha3_256,
+  _chacha20poly1305Decrypt: crypto.chacha20poly1305Decrypt,
+  _chacha20poly1305Encrypt: crypto.chacha20poly1305Encrypt,
 }
