+++ package.json
@@ -27,6 +27,7 @@
     "tape": "^4.11.0"
   },
   "dependencies": {
+    "bech32": "^1.1.4",
     "bip39": "^3.0.2",
     "borc": "^2.1.1",
     "pbkdf2": "^3.0.17"
